Port from Ceylon to Java.
This is as direct a port as I could do (by hand) and still have it
compile and run; I expect to fix massive numbers of warnings in
the next few commits, and there are other obvious improvements that
could conceivably introduce problems and so also merit separate
commits. I also added Maven build configuration so CI can build it.

Fixes #12.