DanceSearchResultsListModel: Explicitly compare using current locale.
That is, compare query to dance names after lowercasing both of them
in the current locale. This is the default, but static analysis tools
complain if the no-argument version of toLowerCase() is used, and it
is indeed clearer that using the user's locale is the intention rather
than just the default.