Specify 'no' locale for OS-name comparisons.
The documentation for String::toLowerCase() says to use Locale.ENGLISH for
locale-insensitive conversions, so we now use it explicitly for testing the
OS name property to see if we're on a Mac. Static analysis tools object when
the no-argument form of toLowerCase() is used, and it's conceivable that
there's some locale where this would break if we didn't specify a locale.