Bump sqlite-jdbc dependency.
The old version (before the original version of this commit; the version
after the Dependabot bump might have been a minimal fix already) has a
security vulnerability, but I don't believe it's possible for this
application to be affected by it.

There is a newer version, but I need to test whether this will work
with it; the other application I wrote that uses it breaks with the
latest versions, but we don't use the JDBC wrapper here that it does.