LaTeXReader: Improve handling of newlines.
We now accept either carriage-return, form-feed, or both in either order as ending a
comment or as ignored when following a LaTeX explicit-newline command, but two
consecutive carriage returns or line feeds are treated as two consecutive newlines.

This also allows us to minimize the number of suppressions of the 'hardcoded
line separator' warning.