Switch nullness annotation provider from IntelliJ to JSpecify
And fix any warnings that are immediately shown. It's possible that the code is now missing
some @Nullable annotations and so the bytecode will contain require-non-null assertion-checks
on values coming from the JDK that are documented to sometimes return null, but I'll do my
best to find and fix any such as I notice them.